added visual regression tests
