+++ package.json
@@ -46,6 +46,7 @@
     "chromedriver": "^2.31.0",
     "glob": "^7.1.1",
     "load-json-file": "^2.0.0",
+    "wdio-visual-regression-service": "^0.8.0",
     "webdriverio": "^4.8.0"
   },
   "peerDependencies": {
