--- src/wdio-terra-service/index.js
@@ -1,3 +1,0 @@
-import TerraLauncher from './TerraLauncher';
-
-module.exports = new TerraLauncher();
