+++ src/wdio/services/index.js
@@ -1,0 +1,5 @@
+import AxeService from './AxeService';
+import TerraService from './TerraService';
+
+module.exports.Axe = new AxeService();
+module.exports.Terra = new TerraService();
