+++ tests/fixtures/compare.html
@@ -7,6 +7,7 @@
     <title>Accessibility Test</title>
   </head>
   <body>
-    <h1 class="test">This text is accessible</h1>
+    <h1 class="test">Hello World</h1/>
+    <button>Press Me</button>
   </body>
 </html>
